Split node commands into separate binaries
