+++ Cargo.toml
@@ -1,6 +1,8 @@
 [workspace]
 members = [
   "cmd/mojave",
+  "cmd/full-node",
+  "cmd/sequencer",
   "crates/block-builder",
   # "crates/full-node",
   "crates/networking",
