+++ cmd/mojave/lib.rs
@@ -1,4 +1,3 @@
-pub mod cli;
 pub mod command;
 pub mod full_node_options;
 pub mod initializer;
@@ -6,6 +5,6 @@ pub mod logging;
 pub mod networks;
 pub mod options;
 pub mod sequencer_options;
-pub(crate) mod version;
+pub mod version;
 
 pub const DEFAULT_DATADIR: &str = "mojave";
