+++ cmd/sequencer/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "mojave-sequencer"
+version = "0.1.0"
+edition = "2021"
+
+[[bin]]
+name = "sequencer"
+path = "src/main.rs"
+
+[dependencies]
+anyhow = { workspace = true }
+clap = { workspace = true, features = ["derive", "env", "string"] }
+tokio = { workspace = true, features = ["full"] }
+tracing = { workspace = true }
+mojave = { path = "../mojave" }
